This version I have commented out the google translate imports to get immediate compatibility on Gcloud from my local. I fixed many errors on the parsing again, mostly with geeralization  and better checking for webpage hits with driver etc.
Also fixed an bad issure where the proxy list would not be recomputed
when it reache3d its end but break instead.
So that should continue indefiniteluy from now on.
i also added in print to show various timing in parsing func. and
surebet checks etc. as well as nu,mber of times the odds dict. changes
(i.e odds across websites) and total no. of surebets found in the
current run.
Made a start to also sending the distribution (i.e three values of the
surebets to place) , given the initial stake to distribute across.